Move ResolvableProfile class to the same file as its parent class
