+++ build-scripts/compile_profiles.py
@@ -3,67 +3,9 @@ from __future__ import print_function
 import argparse
 import sys
 import os.path
-from copy import deepcopy
 from glob import glob
 
 import ssg.build_yaml
-
-
-class ResolvableProfile(ssg.build_yaml.Profile):
-    def __init__(self, * args, ** kwargs):
-        super(ResolvableProfile, self).__init__(* args, ** kwargs)
-        self.resolved = False
-
-    def resolve(self, all_profiles):
-        if self.resolved:
-            return
-
-        resolved_selections = set(self.selected)
-        if self.extends:
-            if self.extends not in all_profiles:
-                msg = (
-                    "Profile {name} extends profile {extended}, but"
-                    "only profiles {known_profiles} are available for resolution."
-                    .format(name=self.id_, extended=self.extends,
-                            profiles=list(all_profiles.keys())))
-                raise RuntimeError(msg)
-            extended_profile = all_profiles[self.extends]
-            extended_profile.resolve(all_profiles)
-
-            extended_selects = set(extended_profile.selected)
-            resolved_selections.update(extended_selects)
-
-            updated_variables = dict(extended_profile.variables)
-            updated_variables.update(self.variables)
-            self.variables = updated_variables
-
-            extended_refinements = deepcopy(extended_profile.refine_rules)
-            updated_refinements = self._subtract_refinements(extended_refinements)
-            updated_refinements.update(self.refine_rules)
-            self.refine_rules = updated_refinements
-
-        for uns in self.unselected:
-            resolved_selections.discard(uns)
-
-        self.unselected = []
-        self.extends = None
-
-        self.selected = sorted(resolved_selections)
-
-        self.resolved = True
-
-    def _subtract_refinements(self, extended_refinements):
-        """
-        Given a dict of rule refinements from the extended profile,
-        "undo" every refinement prefixed with '!' in this profile.
-        """
-        for rule, refinements in list(self.refine_rules.items()):
-            if rule.startswith("!"):
-                for prop, val in refinements:
-                    extended_refinements[rule[1:]].remove((prop, val))
-                del self.refine_rules[rule]
-        return extended_refinements
-
 
 def create_parser():
     parser = argparse.ArgumentParser()
@@ -91,7 +33,7 @@ def make_name_to_profile_mapping(profile_files, env_yaml):
     name_to_profile = {}
     for f in profile_files:
         try:
-            p = ResolvableProfile.from_yaml(f, env_yaml)
+            p = ssg.build_yaml.ResolvableProfile.from_yaml(f, env_yaml)
             name_to_profile[p.id_] = p
         except Exception as exc:
             # The profile is probably doc-incomplete
